btrfs: test if the capability is kept on incremental send
There is a situation where the incremental send can drop the capability
from the receiving side. If you have a file that changed the group, but
the capability was the same of before the group changed, the current
kernel code will only emit a chown, since the capability is the same.

The steps bellow can reproduce the problem.

$ mount /dev/sda fs1
$ mount /dev/sdb fs2

$ touch fs1/foo.bar
$ setcap cap_sys_nice+ep fs1/foo.bar
$ btrfs subvol snap -r fs1 fs1/snap_init
$ btrfs send fs1/snap_init | btrfs receive fs2

$ chgrp adm fs1/foo.bar
$ setcap cap_sys_nice+ep fs1/foo.bar

$ btrfs subvol snap -r fs1 fs1/snap_complete
$ btrfs subvol snap -r fs1 fs1/snap_incremental

$ btrfs send fs1/snap_complete | btrfs receive fs2
$ btrfs send -p fs1/snap_init fs1/snap_incremental | btrfs receive fs2

At this point, a chown was emitted by "btrfs send" since only the group
was changed. This makes the cap_sys_nice capability to be dropped from
fs2/snap_incremental/foo.bar

This test fails on current kernel, the fix was submitted to the btrfs
mailing list titled:

"btrfs: send: Emit file capabilities after chown"

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>